Fix int overflow in KSW2_gg2_SSE memory allocation

+++ ksw2_gg2_sse.c
@@ -33,7 +33,7 @@ int ksw_gg2_sse(void *km, int qlen, const uint8_t *query, int tlen, const uint8_
 	u = (__m128i*)(((size_t)mem + 15) >> 4 << 4); // 16-byte aligned
 	v = u + tlen_, x = v + tlen_, y = x + tlen_, s = y + tlen_;
 	qr = (uint8_t*)kcalloc(km, qlen, 1);
-	mem2 = (uint8_t*)kmalloc(km, ((qlen + tlen - 1) * n_col_ + 1) * 16);
+	mem2 = (uint8_t*)kmalloc(km, ((size_t)(qlen + tlen - 1) * n_col_ + 1) * 16);
 	p = (__m128i*)(((size_t)mem2 + 15) >> 4 << 4);
 	off = (int*)kmalloc(km, (qlen + tlen - 1) * sizeof(int));
 
@@ -66,7 +66,7 @@ int ksw_gg2_sse(void *km, int qlen, const uint8_t *query, int tlen, const uint8_
 		x1_ = _mm_cvtsi32_si128(x1);
 		v1_ = _mm_cvtsi32_si128(v1);
 		st_ = st>>4, en_ = en>>4;
-		pr = p + r * n_col_ - st_;
+		pr = p + (size_t)r * n_col_ - st_;
 		for (t = st_; t <= en_; ++t) {
 			__m128i d, z, a, b, xt1, vt1, ut, tmp;
